I think this is about wrapped up as far as restructuring goes. Hmm.
