+++ src/pack.rs
@@ -1,1 +1,16 @@
+use error::Error;
+use std::process::Command;
 
+pub fn npm_pack(path: &str) -> Result<(), Error> {
+    let pkg_file_path = format!("{}/pkg", path);
+    let output = Command::new("npm")
+        .current_dir(pkg_file_path)
+        .arg("pack")
+        .output()?;
+    if !output.status.success() {
+        let s = String::from_utf8_lossy(&output.stderr);
+        Error::cli("Packaging up your code failed", s)
+    } else {
+        Ok(())
+    }
+}
